requirements.txt
torch==2.7.1
scikit-learn==1.7.1
numpy==2.1.2
transformers==4.53.1
PyYAML==6.0.2
tqdm==4.67.1
openai==1.100.2
python-dotenv==1.1.1